fix plot
